requirements.txt
agent-client-protocol==0.7.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-jinja2==1.6
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asn1crypto==1.5.1
asttokens==3.0.1
asyncpg==0.30.0
attrs==25.4.0
Authlib==1.6.6
backoff==2.2.1
bashlex==0.18
beartype==0.22.9
beautifulsoup4==4.14.3
bidict==0.23.1
binaryornot==0.4.4
bleach==6.3.0
boto3==1.42.17
botocore==1.42.17
browser-use==0.11.2
browser-use-sdk==2.0.12
browsergym-core==0.13.3
bubus==1.5.6
build==1.3.0
CacheControl==0.14.4
cachetools==5.5.2
cdp-use==1.4.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.1
cloudpickle==3.1.2
clr_loader==0.2.9
cmakeast==0.0.18
cobble==0.1.4
comm==0.2.3
contourpy==1.3.3
crashtest==0.4.1
cryptography==46.0.3
cycler==0.12.1
cyclopts==4.4.3
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
dirhash==0.5.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.4
dulwich==0.24.10
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fakeredis==2.33.0
Farama-Notifications==0.0.4
fastapi==0.128.0
fastjsonschema==2.21.2
fastmcp==2.14.1
fastuuid==0.14.0
filelock==3.20.1
findpython==0.7.1
flake8==7.3.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
func_timeout==4.3.5
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
google-cloud-aiplatform==1.130.0
google-cloud-bigquery==3.39.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-genai==1.55.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grep-ast==0.9.0
groq==1.0.0
grpc-google-iam-v1==0.14.3
grpcio==1.67.1
grpcio-status==1.67.1
gymnasium==1.2.3
h11==0.16.0
hf-xet==1.2.0
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-aiohttp==0.1.12
httpx-sse==0.4.3
huggingface_hub==1.2.3
idna==3.11
importlib_metadata==8.7.1
inquirerpy==0.3.4
installer==0.7.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
json_repair==0.54.3
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-kernel-gateway==3.0.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.16
keyring==25.7.0
kiwisolver==1.4.9
kubernetes==33.1.0
lark==1.3.1
libcst==1.5.0
libtmux==0.53.0
linkify-it-py==2.0.3
litellm==1.80.17
lmnr==0.7.25
lupa==2.6
lxml==6.0.2
Mako==1.3.10
mammoth==1.11.0
markdown-it-py==4.0.0
markdownify==1.2.2
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mcp==1.25.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.2.0
more-itertools==10.8.0
msgpack==1.1.2
multidict==6.7.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.0
oauthlib==3.3.1
ollama==0.6.1
openai==2.8.0
openapi-pydantic==0.5.1
openhands==1.9.1
openhands-aci==0.3.2
openhands-agent-server==1.5.2
openhands-ai==1.0.0
openhands-sdk==1.8.2
openhands-tools==1.8.2
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-threading==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-semantic-conventions-ai==0.4.13
orjson==3.11.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathable==0.4.4
pathspec==0.12.1
pathvalidate==3.3.1
pbs-installer==2025.12.17
pdfminer.six==20251228
pexpect==4.9.0
pfzy==0.3.4
pg8000==1.31.5
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.5.1
playwright==1.57.0
poetry==2.2.1
poetry-core==2.2.1
portalocker==2.10.1
posthog==7.4.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase62==1.0.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.16.3
pydub==0.25.1
pyee==13.0.0
pyflakes==3.4.0
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
PyNaCl==1.6.1
pyotp==2.9.0
pyparsing==3.3.1
pypdf==6.5.0
PyPDF2==3.0.1
pyperclip==1.11.0
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-engineio==4.13.0
python-frontmatter==1.1.0
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.21
python-pptx==1.0.2
python-socketio==5.16.0
pythonnet==3.0.5
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qtconsole==5.7.0
QtPy==2.4.3
RapidFuzz==3.14.3
redis==6.4.0
referencing==0.36.2
regex==2025.11.3
reportlab==4.4.7
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
scantree==0.0.4
scipy==1.16.3
scramp==1.4.6
screeninfo==0.8.1
SecretStorage==3.5.0
selenium==4.39.0
Send2Trash==1.8.3
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
SpeechRecognition==3.14.4
SQLAlchemy==2.0.45
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.48.0
streamingjson==0.0.5
tenacity==9.1.2
termcolor==3.3.0
terminado==0.18.1
textual==7.3.0
textual-autocomplete==4.0.6
textual-serve==1.1.3
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
tom-swe==1.0.3
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
tree-sitter==0.24.0
tree-sitter-c-sharp==0.23.1
tree-sitter-embedded-template==0.25.0
tree-sitter-language-pack==0.7.3
tree-sitter-yaml==0.7.2
trio==0.32.0
trio-websocket==0.12.2
trove-classifiers==2025.12.1.14
typer==0.21.0
typer-slim==0.21.0
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.2
uuid7==0.1.0
uvicorn==0.40.0
virtualenv==20.35.4
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
whatthepatch==1.0.7
widgetsnbextension==4.0.15
wrapt==1.17.3
wsproto==1.3.2
xlrd==2.0.2
xlsxwriter==3.2.9
yarl==1.22.0
youtube-transcript-api==1.2.3
zipp==3.23.0
zope.interface==7.2
zstandard==0.25.0
